# openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/rewards/{id}:
    get:
      operationId: AdminRewardsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados da recompensa
        '404':
          description: Recompensa não encontrada
      security: *ref_4
      summary: Retorna os detalhes de uma recompensa
      tags: *ref_5
    patch:
      operationId: AdminRewardsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRewardDto'
      responses:
        '200':
          description: Recompensa atualizada
        '404':
          description: Recompensa não encontrada
      security: *ref_4
      summary: Atualiza uma recompensa existente
      tags: *ref_5
    delete:
      operationId: AdminRewardsController_softDelete
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Recompensa removida
      security: *ref_4
      summary: Soft delete de uma recompensa
      tags: *ref_5
  /rewards:
    get:
      operationId: RewardsController_listAvailable
      parameters: []
      responses:
        '200':
          description: Vitrine de recompensas disponíveis
      security: &ref_6
        - bearer: []
      summary: Lista recompensas disponíveis (ativas e com estoque)
      tags: &ref_7
        - rewards
  /rewards/{id}:
    get:
      operationId: RewardsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados da recompensa
        '404':
          description: Recompensa não encontrada ou indisponível
      security: *ref_6
      summary: Retorna os detalhes de uma recompensa disponível
      tags: *ref_7
  /rewards/{id}/redeem:
    post:
      operationId: RewardsController_redeem
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '201':
          description: Resgate solicitado com sucesso, pontos debitados
        '400':
          description: Saldo insuficiente, sem estoque ou cooldown ativo
        '404':
          description: Recompensa não encontrada
      security: *ref_6
      summary: Solicita o resgate de uma recompensa
      tags: *ref_7
...
  /admin/redemptions/{id}:
    get:
      operationId: AdminRedemptionsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados do resgate
        '404':
          description: Resgate não encontrado
      security: *ref_10
      summary: Retorna os detalhes de um resgate específico
      tags: *ref_11
...
  /redemptions:
    get:
      operationId: RedemptionsController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            example: 1
            type: number
        - name: limit
          required: false
          in: query
          schema:
            example: 20
            type: number
      responses:
        '200':
          description: Histórico paginado de resgates do membro
      security: &ref_12
        - bearer: []
      summary: Lista o histórico de resgates do membro autenticado
      tags: &ref_13
        - redemptions
  /redemptions/{id}:
    get:
      operationId: RedemptionsController_findById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Dados do resgate
        '404':
          description: Resgate não encontrado
      security: *ref_12
      summary: Retorna os detalhes de um resgate específico do membro
      tags: *ref_13
...
info:
  title: domy-api
  description: API do app doméstico gamificado
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateRewardDto:
      type: object
      properties: {}
...